09_SingleAgent/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
openai>=2.0.0
chromadb==1.5.8
python-dotenv==1.0.1
pydantic>=2.0.0
httpx>=0.27.0
langchain>=0.3.0
langchain-openai>=0.2.0
langchain-chroma>=0.1.0
rank_bm25>=0.2.2
kiwipiepy>=0.17.0
flashrank>=0.2.10
python-jose[cryptography]>=3.3.0
python-multipart>=0.0.9
tavily-python>=0.3.0
mysql-connector-python==8.3.0
